requirements.txt
# CK3 Character Manager Dependencies
Pillow>=10.0.0
